Inline D3D macros into files
